requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
flask-swagger-ui==4.11.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
bcrypt==4.1.2
PyJWT==2.8.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-huggingface==0.3.1
langchain-ollama==0.3.8
langchain-text-splitters==0.3.11
chromadb==0.4.24
pypdf==4.2.0
boto3==1.34.91
botocore==1.34.91
tenacity==8.2.3
python-dotenv==1.0.1
pytest==8.1.1
pytest-mock==3.14.0
sentence-transformers==5.1.0
